chore: move getUniqueId to utils
